server/models/requirements.txt
pytesseract>=0.3.10
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
pdf2image>=1.16.3
python-magic-bin>=0.4.14 ; sys_platform == 'win32'
python-magic>=0.4.27 ; sys_platform != 'win32'
watchdog>=3.0.0
pytest>=7.4.0
pytest-cov>=4.1.0
structlog>=23.0.0
scikit-image>=0.21.0
imutils>=0.5.4
pywin32>=306 ; sys_platform == 'win32' and platform_python_implementation == 'CPython'
transformers>=4.35.0
torch>=2.0.0
datasets>=2.14.0
peft>=0.6.0
trl>=0.7.0
scikit-learn>=1.3.0
sentencepiece>=0.1.99
protobuf>=3.20.0
accelerate>=0.24.0